Refactor README.md to remove outdated sections and improve clarity. Update crew.py to replace deprecated tools with YFinanceMACDTool and LongBridgeMACDTool for enhanced data retrieval. Modify main.py to streamline execution and improve logging. Update agents.yaml and tasks.yaml for clearer role descriptions and task dependencies, ensuring better alignment with trading strategies.
