+++ src/macd_trader/crew.py
@@ -3,33 +3,13 @@ from crewai.project import CrewBase, agent, crew, task
 
 from src.llm import LLM
 from src.macd_trader.tools.notification_tools import WechatNotificationTool
-from src.macd_trader.tools.stock_data_tools import StockDataTools
-from src.macd_trader.tools.trading_tools import LBQuoteMACDTool
-# Assuming you might use Deepseek directly or another OpenAI-compatible API
-# You might need to adjust the base_url and api_key parameters
-# Example for OpenAI compatible endpoint (like local LLM server or other providers)
-# llm = ChatOpenAI(
-#     model = "deepseek-coder", # Or your specific model
-#     base_url = "YOUR_API_ENDPOINT", # e.g., http://localhost:11434/v1
-#     api_key = os.getenv("DEEPSEEK_API_KEY", "dummy_key") # API key might not be needed for local models
-# )
+from src.macd_trader.tools.yfinance_tools import YFinanceMACDTool
+from src.macd_trader.tools.longbridge_tools import LongBridgeMACDTool
 
-# If Deepseek has its own Langchain integration class:
-# from langchain_community.chat_models import ChatDeepseek # Fictional example
-# llm = ChatDeepseek(model="deepseek-coder", api_key=os.getenv("DEEPSEEK_API_KEY"))
-
-# Fallback: Using a standard OpenAI compatible model for structure
-# Replace this with your actual Deepseek LLM setup
-# deepseek_llm = LLM.deepseek()
-
-
-# Import your custom tools
-
-
-# Instantiate tools
 notification_tool = WechatNotificationTool()
-yf_fetcher_tool = StockDataTools()
-lb_fetcher_tool = LBQuoteMACDTool()
+yf_fetcher_tool = YFinanceMACDTool()
+lb_fetcher_tool = LongBridgeMACDTool()
+llm = LLM.deepseek()
 
 
 @CrewBase
@@ -39,18 +19,12 @@ class TradingCrew:
     agents_config = "config/agents.yaml"
     tasks_config = "config/tasks.yaml"
 
-    def __init__(self) -> None:
-        # If using a non-OpenAI model compatible with Langchain ChatModel interface:
-        self.llm = LLM.deepseek()  # Use the llm defined above
-        # If the Deepseek model requires direct SDK use within agents/tools,
-        # you might not define a global LLM here, or pass it differently.
-
     @agent
     def data_fetcher(self) -> Agent:
         return Agent(
             config=self.agents_config["data_fetcher"],
             tools=[lb_fetcher_tool],
-            llm=self.llm,  # Uncomment if you want to explicitly pass the LLM
+            llm=llm,
             verbose=True,
         )
 
@@ -58,8 +32,7 @@ class TradingCrew:
     def trading_strategist(self) -> Agent:
         return Agent(
             config=self.agents_config["trading_strategist"],
-            # No specific tools needed here as logic is in the task description
-            llm=self.llm,
+            llm=llm,
             verbose=True,
         )
 
@@ -67,7 +40,7 @@ class TradingCrew:
     def investment_advisor(self) -> Agent:
         return Agent(
             config=self.agents_config["investment_advisor"],
-            llm=self.llm,
+            llm=llm,
             tools=[notification_tool],
             verbose=True,
         )
@@ -81,17 +54,6 @@ class TradingCrew:
     #         verbose=True,
     #     )
 
-    # @agent
-    # def notifier(self) -> Agent:
-    #     return Agent(
-    #         config=self.agents_config["notifier"],
-    #         tools=[
-    #             WechatNotificationTool()
-    #         ],
-    #         llm=self.llm,
-    #         verbose=True,
-    #     )
-
     # --- Tasks --- #
 
     @task
@@ -99,19 +61,22 @@ class TradingCrew:
         return Task(
             config=self.tasks_config["fetch_data_task"],
             agent=self.data_fetcher(),
-            # You might pass expected output format details here if needed
         )
 
     @task
     def analyze_macd_task(self) -> Task:
-        # This task requires a more complex implementation than just description
-        # It needs to parse the MACD data and apply the crossover logic.
-        # CrewAI tasks can execute code directly or rely on agent capabilities/tools.
-        # Let's define it to depend on the fetch_data_task context.
         return Task(
             config=self.tasks_config["analyze_macd_task"],
             agent=self.trading_strategist(),
-            context=[self.fetch_data_task()],  # Pass output from previous task
+            context=[self.fetch_data_task()],
+        )
+
+    @task
+    def generate_advice_task(self) -> Task:
+        return Task(
+            config=self.tasks_config["generate_advice_task"],
+            agent=self.investment_advisor(),
+            context=[self.analyze_macd_task()],
         )
 
     # @task
@@ -119,47 +84,15 @@ class TradingCrew:
     #     return Task(
     #         config=self.tasks_config["execute_trade_task"],
     #         agent=self.trader(),
-    #         context=[self.analyze_macd_task()],  # Pass decision from strategist
+    #         context=[self.analyze_macd_task()],
     #     )
-
-    # @task
-    # def send_notification_task(self) -> Task:
-    #     return Task(
-    #         config=self.tasks_config["send_notification_task"],
-    #         agent=self.notifier(),
-    #         context=[self.analyze_macd_task()],  # Pass trade execution result
-    #     )
-
-    @task
-    def generate_advice_task(self) -> Task:
-        return Task(
-            config=self.tasks_config["generate_advice_task"],
-            agent=self.investment_advisor(),
-            context=[self.analyze_macd_task()],  # 使用MACD分析结果
-        )
 
     @crew
     def crew(self) -> Crew:
         """Creates the Trading Crew."""
         return Crew(
-            manager_llm=self.llm,
-            function_calling_llm=self.llm,
-            chat_llm=self.llm,
-            agents=[
-                self.data_fetcher(),
-                self.trading_strategist(),
-                # self.trader()
-                # self.notifier(),
-                self.investment_advisor(),
-            ],
-            tasks=[
-                self.fetch_data_task(),
-                self.analyze_macd_task(),
-                # self.execute_trade_task(),
-                # self.send_notification_task(),
-                self.generate_advice_task(),
-            ],
-            process=Process.sequential,  # Tasks will run one after another
-            verbose=False,  # Set verbosity level (0, 1, or 2)
-            # memory=True # Enable memory for conversational context if needed
+            agents=self.agents,
+            tasks=self.tasks,
+            process=Process.sequential,
+            verbose=False,
         )
