+++ src/macd_trader/main.py
@@ -5,15 +5,9 @@ import sys
 from src.macd_trader.crew import TradingCrew
 
 # Add the project root to the Python path
-# This allows imports like `from my_project.crew import TradingCrew`
 project_root = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
 sys.path.insert(0, os.path.dirname(project_root))
 
-# Load environment variables from .env file
-
-# Import the crew after setting the path and loading env vars
-
-# Configure logging
 logging.basicConfig(
     level=logging.INFO, format="%(asctime)s - %(name)s - %(levelname)s - %(message)s"
 )
@@ -21,9 +15,7 @@ logger = logging.getLogger(__name__)
 
 
 def run():
-    """Initializes and runs the Trading Crew."""
-    # Get inputs from environment variables or provide defaults
-    stock_ticker = os.getenv("TARGET_STOCK", "NVDA.US")  # Default to AAPL if not set
+    stock_ticker = os.getenv("TARGET_STOCK", "NVDA.US")  # Default to NVDA.US if not set
     trade_quantity = int(os.getenv("TRADE_QUANTITY", "1"))  # Default to 10 if not set
 
     inputs = {"stock_ticker": stock_ticker, "quantity": trade_quantity}
@@ -33,12 +25,7 @@ def run():
     )
 
     try:
-        # Initialize the crew
         trading_crew = TradingCrew()
-
-        # Kick off the crew's process
-        # The inputs dictionary variables ({stock_ticker}, {quantity}) will be automatically
-        # interpolated into the agent goals and task descriptions where defined.
         result = trading_crew.crew().kickoff(inputs=inputs)
 
         logger.info("\n\nTrading Crew finished execution.")
